Move trip update providers to own folder
