+++ src/providers/tripUpdates/TampereProvider.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosRequestConfig } from 'axios';
-import config from '../config/config';
-import { storeTripUpdateFeed } from '../lib/tampereJourneysApiFeedProcessor';
-import { DataProvider } from '../providers';
+import config from '../../config/config';
+import { storeTripUpdateFeed } from '../../lib/tampereJourneysApiFeedProcessor';
+import { DataProvider } from '.';
 
 /**
  * Tampere Journeys api provider
